chore(refactor): making transfer actor part of client
