+++ Cargo.toml
@@ -53,7 +53,6 @@ serde_json = "1.0.9"
 tempfile = "3.1.0"
 
 [features]
-mock-network = ["simulated-payouts"]
 simulated-payouts = ["safe-nd/simulated-payouts", "safe-transfers/simulated-payouts"]
 testing = ["simulated-payouts"]
