+++ src/crypto.rs
@@ -99,10 +99,17 @@ pub mod shared_box {
     }
 
     /// Generate new random public/secret keypair.
-    pub fn gen_keypair() -> (threshold_crypto::PublicKey, SecretKey) {
+    pub fn gen_keypair() -> (SecretKey, threshold_crypto::PublicKey) {
         let sk = threshold_crypto::SecretKey::random();
         let pk = sk.public_key();
-        (pk, SecretKey::new(sk))
+        (SecretKey::new(sk), pk)
+    }
+
+    /// Generate new random public/secret keypair.
+    pub fn gen_bls_keypair() -> (threshold_crypto::SecretKey, threshold_crypto::PublicKey) {
+        let sk = threshold_crypto::SecretKey::random();
+        let pk = sk.public_key();
+        (sk, pk)
     }
 
     impl Deref for SecretKey {
