+++ src/lib.rs
@@ -36,12 +36,16 @@
 // Export public core interface.
 
 pub use self::client::{
-    map_info, test_create_balance, Client, ClientTransferValidator, MapInfo, TransferActor,
+    map_info,
+    // test_create_balance,
+    Client,
+    MapInfo,
+    // transfer_actor
 };
 pub use self::config_handler::config_dir;
 pub use self::connection_manager::ConnectionManager;
 pub use self::errors::CoreError;
-pub use self::network_event::{NetworkEvent, NetworkRx, NetworkTx};
+// pub use self::network_event::{NetworkEvent, NetworkRx, NetworkTx};
 pub use self::utils::logging;
 pub use quic_p2p::Config as QuicP2pConfig;
 
@@ -56,7 +60,7 @@ pub mod utils;
 
 mod connection_manager;
 mod errors;
-mod network_event;
+// mod network_event;
 
 /// `Map` type tag for a directory.
 pub const DIR_TAG: u64 = 15_000;
