+++ src/utils/test_utils/mod.rs
@@ -6,16 +6,15 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::client::{Client, COST_OF_PUT};
-use crate::network_event::{NetworkEvent, NetworkTx};
-use crate::utils::{self};
-use crate::CoreError;
-use futures::{channel::mpsc, future::Future};
-use log::trace;
-use safe_nd::{AppFullId, ClientFullId, ClientPublicId, Keypair, Money};
-use std::fmt::Debug;
+use crate::client::{COST_OF_PUT};
 
-use tokio::stream::StreamExt;
+
+
+
+use safe_nd::{ClientFullId, Keypair, Money};
+
+
+
 use unwrap::unwrap;
 
 /// Generates a random BLS secret and public keypair.
@@ -28,84 +27,6 @@ pub fn gen_bls_keypair() -> Keypair {
 pub fn gen_client_id() -> ClientFullId {
     let mut rng = rand::thread_rng();
     ClientFullId::new_bls(&mut rng)
-}
-
-/// Generates a random app full ID.
-pub fn gen_app_id(client_public_id: ClientPublicId) -> AppFullId {
-    let mut rng = rand::thread_rng();
-    AppFullId::new_bls(&mut rng, client_public_id)
-}
-
-/// Create random registered client. Use this to
-/// create a `CoreClient` automatically and randomly.
-pub fn random_client() -> Result<CoreClient, CoreError>
-where
-{
-    // FIXME: in stage 1 disconnection is a natural event, so instead of panicking we
-    // just print it out.
-    let on_network_event = |net_event| trace!("Unexpected NetworkEvent occurred: {:?}", net_event);
-    let client_creator = |net_tx| {
-        let acc_locator = utils::generate_random_string(10);
-        let acc_password = utils::generate_random_string(10);
-
-        // blocking on the thread here as part of tests. Core client construction is async
-        futures::executor::block_on(CoreClient::new(&acc_locator, &acc_password, net_tx))
-    };
-    match setup_client_with_net_obs(&(), client_creator, on_network_event) {
-        Ok((_receiver, client)) => Ok(client),
-        Err(error) => Err(error),
-    }
-}
-
-/// Helper to create a client
-/// Useful when we need  to supply credentials explicitly or when Client is to be constructed as
-/// unregistered or as a result of successful login. Use this to create Client
-/// manually.
-pub fn setup_client<Create, A, C, F>(context: &A, c: Create) -> Result<C, CoreError>
-where
-    Create: FnOnce(NetworkTx) -> Result<C, F>,
-    A: 'static,
-    C: Client,
-    F: Debug,
-{
-    // FIXME: in stage 1 disconnection is a natural event, so instead of panicking we
-    // just print it out.
-    let n = |net_event| trace!("Unexpected NetworkEvent occurred: {:?}", net_event);
-    match setup_client_with_net_obs(context, c, n) {
-        Ok((_receiver, client)) => Ok(client),
-        Err(error) => Err(error),
-    }
-}
-
-/// Helper to create a client and setup network event listener.
-/// Useful when we need
-/// to supply credentials explicitly or when Client is to be constructed as
-/// unregistered or as a result of successful login. Use this to create Client
-/// manually.
-pub fn setup_client_with_net_obs<Create, NetObs, A, C, F>(
-    _context: &A,
-    client_creator: Create,
-    mut n: NetObs,
-) -> Result<(Box<dyn Future<Output = ()> + 'static + Send>, C), CoreError>
-where
-    Create: FnOnce(NetworkTx) -> Result<C, F>,
-    NetObs: FnMut(NetworkEvent) + 'static + Send,
-    A: 'static,
-    C: Client,
-
-    F: Debug,
-{
-    let (net_tx, mut net_rx) = mpsc::unbounded();
-    let client = unwrap!(client_creator(net_tx));
-
-    let net_fut = async move {
-        while let Some(msg) = net_rx.next().await {
-            n(msg);
-        }
-    };
-
-    // net fut returned in order to keep it alive.
-    Ok((Box::new(net_fut), client))
 }
 
 /// Helper function to calculate the total cost of expenditure by adding number of mutations and
